docs: 프로젝트명 변경, 모듈별 디렉토리 추가
- 프로젝트명 변경
- 모듈별 디렉토리 추가
