--- src/main/java/com/megazone/ERPSystem_phase2/ErpSystemPhase2Application.java
@@ -1,13 +1,0 @@
-package com.megazone.ERPSystem_phase2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ErpSystemPhase2Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ErpSystemPhase2Application.class, args);
-	}
-
-}
