--- src/main/java/com/megazone/ERPSystem_phase2/financial/model/basic_information_management/company_registration/BusinessItem.java
@@ -1,23 +1,0 @@
-package com.megazone.ERPSystem_phase2.financial.model.basic_information_management.company_registration;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * 회사 및 거래처 등록의 업체종목
- */
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class BusinessItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false)
-    private Long id; // 고유 식별자
-
-    @Column(nullable = false) // 종목이름
-    private String name;
-}
