--- src/main/java/com/megazone/ERPSystem_phase2/financial/model/enums/EntityType.java
@@ -1,10 +1,0 @@
-package com.megazone.ERPSystem_phase2.financial.model.enums;
-
-/**
- * EntityType enum 클래스
- * 법인과 개인을 구분
- */
-public enum EntityType {
-    INCORPORATED, // 법인임
-    INDIVIDUAL    // 개인임
-}
