--- src/main/java/com/megazone/ERPSystem_phase2/integrated_systems/controller/HomeController.java
@@ -1,16 +1,0 @@
-package com.megazone.ERPSystem_phase2.integrated_systems.controller;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String greeting() {
-        return "index";
-    }
-
-}
