+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/ErpSystemPhase2BackendApplication.java
@@ -1,0 +1,13 @@
+package com.megazone.ERPSystem_phase2_Backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ErpSystemPhase2BackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ErpSystemPhase2BackendApplication.class, args);
+	}
+
+}
