+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/financial/model/enums/EntityType.java
@@ -1,0 +1,10 @@
+package com.megazone.ERPSystem_phase2_Backend.financial.model.enums;
+
+/**
+ * EntityType enum 클래스
+ * 법인과 개인을 구분
+ */
+public enum EntityType {
+    INCORPORATED, // 법인임
+    INDIVIDUAL    // 개인임
+}
