+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/integrated_systems/controller/HomeController.java
@@ -1,0 +1,15 @@
+package com.megazone.ERPSystem_phase2_Backend.integrated_systems.controller;
+
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/")
+    public String greeting() {
+        return "index";
+    }
+
+}
