+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: ERPSystem-phase2
+    name: ERPSystem-phase2-Backend
   datasource:
     url: jdbc:mysql://localhost:3306/erp_test?useSSL=false&serverTimezone=UTC
     username: root
